create git mixin
