+++ lib/msf/core/exploit/mixins.rb
@@ -67,6 +67,7 @@ require 'msf/core/exploit/afp'
 require 'msf/core/exploit/realport'
 require 'msf/core/exploit/sip'
 require 'msf/core/exploit/tincd'
+require 'msf/core/exploit/git'
 
 # Telephony
 require 'msf/core/exploit/dialup'
